[dv] Use `id_done` to accurately track instruction monitor
This refactors `instr_vif` to use `rvfi_id_done` instead of
`instr_new_id` to track when a new instruction appears in the ID stage.
This interface and signal are only used to keep track of instruction
fetch errors by using the aforementioned valid signal and checking
whether the `rvfi_order_id` has changed. However, if an instruction
fetch error is consecutive to another error, `instr_new_id` will be
gated, which leads us to miss the `fetch_err` in the verification.

This will fix failing `riscv_mem_error_test`